feat: #1661 add market-expires to market info components
