+++ apps/trading/pages/markets/trade-grid.tsx
@@ -1,8 +1,5 @@
 import { DealTicketContainer } from '@vegaprotocol/deal-ticket';
-import {
-  getMarketExpiryDate,
-  MarketInfoContainer,
-} from '@vegaprotocol/market-info';
+import { MarketInfoContainer, getExpiryDate } from '@vegaprotocol/market-info';
 import { OrderbookContainer } from '@vegaprotocol/market-depth';
 import { OrderListContainer } from '@vegaprotocol/orders';
 import { FillsContainer } from '@vegaprotocol/fills';
@@ -23,7 +20,7 @@ import {
   ButtonLink,
   Link,
 } from '@vegaprotocol/ui-toolkit';
-import { getDateFormat, t } from '@vegaprotocol/react-helpers';
+import { t } from '@vegaprotocol/react-helpers';
 import { useAssetDetailsDialogStore } from '@vegaprotocol/assets';
 import { useEnvironment } from '@vegaprotocol/environment';
 import { Header, HeaderStat } from '../../components/header';
@@ -59,17 +56,7 @@ type ExpiryLabelProps = {
 };
 
 const ExpiryLabel = ({ market }: ExpiryLabelProps) => {
-  const closeDate = getMarketExpiryDate(
-    market.tradableInstrument.instrument.metadata.tags
-  );
-  let content = null;
-  if (!closeDate) {
-    content = t('Not time-based');
-  } else {
-    const isExpired = Date.now() - closeDate.valueOf() > 0;
-    const expiryDate = getDateFormat().format(closeDate);
-    content = `${isExpired ? `${t('Expired')} ` : ''} ${expiryDate}`;
-  }
+  const content = getExpiryDate(market);
   return <div data-testid="trading-expiry">{content}</div>;
 };
