+++ libs/market-info/src/components/market-expires/index.ts
@@ -1,2 +1,1 @@
 export * from './market-expires';
-export * from './market-info';
